+++ src/main/java/com/backend/productprices/domain/productprice/entity/Product.java
@@ -1,4 +1,4 @@
 package com.backend.productprices.domain.productprice.entity;
 
-public record Product(Integer id, Integer brandId) {
+public record Product(Long id, Long brandId) {
 }
